requirements.txt
cloudscraper==1.2.71
web3==7.6.0
python-dotenv==1.0.1
